% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioFile.R
\name{audioFile}
\alias{audioFile}
\title{Select an existing audio file to act as a dummy recording}
\usage{
audioFile(time)
}
\arguments{
\item{time}{time to generate temperature for (timestamp)}
}
\description{
Select an existing audio file to act as a dummy recording
}
