% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{tempHumidity}
\alias{tempHumidity}
\title{Generate dummy temperature and humidity}
\usage{
tempHumidity(time)
}
\arguments{
\item{time}{time to generate temperature for (timestamp)}
}
\description{
Generates a very rough temperature based on current time
}
